- Re-integrate Grizzly 2.2.3.  Changes made to EjbWSAdapter.java to resolve cts smoke test issues.
git-svn-id: https://svn.java.net/svn/glassfish~svn/trunk/main@53038 6f3ba3e3-413c-0410-a8aa-90bee3ab43b5

+++ appserver/web/admin/src/main/java/org/glassfish/web/admin/monitor/HttpServiceStatsProvider.java
@@ -349,7 +349,7 @@ public class HttpServiceStatsProvider implements PostConstruct {
 
     @ProbeListener("glassfish:web:http-service:dataSentEvent")
     public void dataSentEvent(
-        @ProbeParam("size") int size) {
+        @ProbeParam("size") long size) {
         countBytesTransmitted.increment(size);
     }
